bot/requirements.txt
aiogram>=3.0.0
aiogram>=3.4.1
aiohttp
aiohttp>=3.9.3
google-generativeai==0.8.3
google-generativeai>=0.8.3
posthog==3.4.2
pydantic>=2.6.4
python-dotenv
python-dotenv==1.0.1
reportlab==4.1.0
supabase
supabase>=2.3.4
